Make the quotes carousel a separate component from Quote
