+++ .gitignore
@@ -12,6 +12,7 @@ pids
 *.pid
 *.seed
 *.pid.lock
+cms-cache.js
 
 # Directory for instrumented libs generated by jscoverage/JSCover
 lib-cov
