+++ components/QuotesCarousel.module.css
